# api/swagger/swagger.yaml
basePath: /
host: localhost:8888
info:
  contact:
    email: kartashov_egor96@mail.ru
    name: mail
    url: /
  description: eShopCDN
  license:
    name: MIT
    url: http://www.apache.org/licenses/MIT.html
  termsOfService: /
  title: eShopCDN Swagger Doc
  version: "1.0"
paths:
  /api/v1/cdn/delete/{id}:
    delete:
      consumes:
      - application/json
      description: delete file by name
      parameters:
      - description: "123"
        in: header
        name: x-api-key
        required: true
        type: string
      - description: Filenam
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Allow delete file
      tags:
      - Files Controller
...
schemes:
- http
swagger: "2.0"